renamed api calls directory to http
